+++ code/packages/ruby/board_vm/ext/board_vm_native/src/lib.rs
@@ -2,14 +2,14 @@ use std::ffi::{c_char, c_int, c_long, c_void};
 use std::ptr;
 use std::slice;
 
-use board_vm_host::PwmWriteProgram;
 use board_vm_host::{
-    BlinkProgram, GpioHandleCloseProgram, GpioHandleReadProgram, GpioHandleWriteProgram,
-    GpioOpenProgram, GpioReadProgram, GpioWriteProgram, LedMatrixFrameProgram, TimeNowProgram,
-    TimeSleepMsProgram, BLINK_MODULE_LEN, GPIO_HANDLE_CLOSE_MODULE_LEN,
-    GPIO_HANDLE_READ_MODULE_LEN, GPIO_HANDLE_WRITE_MODULE_LEN, GPIO_OPEN_MODULE_LEN,
-    GPIO_READ_MODULE_LEN, GPIO_WRITE_MODULE_LEN, LED_MATRIX_FRAME_MODULE_LEN, PWM_WRITE_MODULE_LEN,
-    TIME_NOW_MODULE_LEN, TIME_SLEEP_MS_MODULE_LEN,
+    AdcReadProgram, BlinkProgram, GpioHandleCloseProgram, GpioHandleReadProgram,
+    GpioHandleWriteProgram, GpioOpenProgram, GpioReadProgram, GpioWriteProgram,
+    LedMatrixFrameProgram, PwmWriteProgram, TimeNowProgram, TimeSleepMsProgram,
+    ADC_READ_MODULE_LEN, BLINK_MODULE_LEN, GPIO_HANDLE_CLOSE_MODULE_LEN, GPIO_HANDLE_READ_MODULE_LEN,
+    GPIO_HANDLE_WRITE_MODULE_LEN, GPIO_OPEN_MODULE_LEN, GPIO_READ_MODULE_LEN,
+    GPIO_WRITE_MODULE_LEN, LED_MATRIX_FRAME_MODULE_LEN, PWM_WRITE_MODULE_LEN, TIME_NOW_MODULE_LEN,
+    TIME_SLEEP_MS_MODULE_LEN,
 };
 use board_vm_language_core::{
     bluetooth_backend_open_plan as core_bluetooth_backend_open_plan,
@@ -18,7 +18,7 @@ use board_vm_language_core::{
     build_gpio_handle_read_module, build_gpio_handle_write_module, build_gpio_open_module,
     build_gpio_read_module, build_gpio_write_module, build_hello_wire_frame,
     build_led_matrix_frame_module, build_program_begin_wire_frame, build_program_chunk_wire_frame,
-    build_program_end_wire_frame, build_pwm_write_module, build_raw_module,
+    build_program_end_wire_frame, build_pwm_write_module, build_adc_read_module, build_raw_module,
     build_run_background_wire_frame, build_run_wire_frame, build_stop_wire_frame,
     build_store_program_wire_frame, build_time_now_module, build_time_sleep_ms_module,
     capability_board_metadata, capability_bytecode_callable, capability_flag_names,
@@ -227,6 +227,19 @@ extern "C" fn session_pwm_write_module(
 
     let module = build_pwm_write_module_value(pin, duty, max_stack)
         .unwrap_or_else(|error| raise_core_error("pwm_write_module", error));
+    ruby_bridge::bytes_to_rb(&module)
+}
+
+extern "C" fn session_adc_read_module(
+    _self_val: VALUE,
+    pin_val: VALUE,
+    max_stack_val: VALUE,
+) -> VALUE {
+    let pin = rb_u8(pin_val, "pin");
+    let max_stack = rb_u8(max_stack_val, "max_stack");
+
+    let module = build_adc_read_module_value(pin, max_stack)
+        .unwrap_or_else(|error| raise_core_error("adc_read_module", error));
     ruby_bridge::bytes_to_rb(&module)
 }
 
@@ -1390,6 +1403,13 @@ fn build_pwm_write_module_value(
     Ok(module)
 }
 
+fn build_adc_read_module_value(pin: u8, max_stack: u8) -> Result<Vec<u8>, LanguageCoreError> {
+    let mut module = vec![0; ADC_READ_MODULE_LEN];
+    let len = build_adc_read_module(AdcReadProgram { pin, max_stack }, &mut module)?;
+    module.truncate(len);
+    Ok(module)
+}
+
 fn build_raw_module_value(
     flags: u8,
     max_stack: u8,
@@ -1692,6 +1712,12 @@ pub extern "C" fn Init_board_vm_native() {
         "pwm_write_module",
         session_pwm_write_module as *const c_void,
         3,
+    );
+    ruby_bridge::define_method_raw(
+        session_class,
+        "adc_read_module",
+        session_adc_read_module as *const c_void,
+        2,
     );
     ruby_bridge::define_method_raw(
         session_class,